Refactor TypeScript
- Add index files.
- Remove redundant import.
- Fix copyright header.
- Simplify constructors.
- Add `readonly` modifiers where relevant.
