+++ src/DependabotHelper/scripts/ts/Client/GitHubClient.ts
@@ -1,10 +1,7 @@
 // Copyright (c) Martin Costello, 2022. All rights reserved.
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
-import { ApiError } from '../Models/ApiError';
-import { RateLimits } from '../Models/RateLimits';
-import { Repository } from '../Models/Repository';
-import { RepositoryPullRequests } from '../Models/RepositoryPullRequests';
+import { ApiError, RateLimits, Repository, RepositoryPullRequests } from '../Models/index';
 
 export class GitHubClient {
     readonly rateLimits: RateLimits;
