+++ src/DependabotHelper/scripts/ts/Client/index.ts
@@ -1,0 +1,5 @@
+// Copyright (c) Martin Costello, 2022. All rights reserved.
+// Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
+
+export { GitHubClient } from './GitHubClient';
+export { StorageClient } from './StorageClient';
