+++ src/DependabotHelper/scripts/ts/Models/index.ts
@@ -1,0 +1,12 @@
+// Copyright (c) Martin Costello, 2022. All rights reserved.
+// Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
+
+export { ApiError } from './ApiError';
+export { ChecksStatus } from './ChecksStatus';
+export { Owner } from './Owner';
+export { OwnerRepositories } from './OwnerRepositories';
+export { PullRequest } from './PullRequest';
+export { RateLimits } from './RateLimits';
+export { Repository } from './Repository';
+export { RepositoryPullRequests } from './RepositoryPullRequests';
+export { UserProfile } from './UserProfile';
