+++ src/DependabotHelper/scripts/ts/View/App.ts
@@ -1,8 +1,7 @@
 // Copyright (c) Martin Costello, 2022. All rights reserved.
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
-import { GitHubClient } from '../Client/GitHubClient';
-import { StorageClient } from '../Client/StorageClient';
+import { GitHubClient, StorageClient } from '../Client/index';
 import { Analytics } from './Analytics';
 import { Configuration } from './Configuration';
 import { ErrorsElement } from './ErrorsElement';
