+++ src/DependabotHelper/scripts/ts/View/Elements.ts
@@ -1,7 +1,8 @@
 // Copyright (c) Martin Costello, 2022. All rights reserved.
+// Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
 export class Elements {
-    static hidden = 'd-none';
+    static readonly hidden = 'd-none';
 
     static disable(element: Element): void {
         element.setAttribute('disabled', '');
