+++ src/DependabotHelper/scripts/ts/View/Manage.ts
@@ -1,9 +1,8 @@
 // Copyright (c) Martin Costello, 2022. All rights reserved.
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
-import { GitHubClient } from '../Client/GitHubClient';
-import { StorageClient } from '../Client/StorageClient';
-import { RepositoryPullRequests } from '../Models/RepositoryPullRequests';
+import { GitHubClient, StorageClient } from '../Client/index';
+import { RepositoryPullRequests } from '../Models/index';
 import { Elements } from './Elements';
 import { ErrorsElement } from './ErrorsElement';
 import { OwnerElement } from './OwnerElement';
