+++ src/DependabotHelper/scripts/ts/View/PullRequestsElement.ts
@@ -1,19 +1,17 @@
 // Copyright (c) Martin Costello, 2022. All rights reserved.
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
-import { PullRequest } from '../Models/PullRequest';
+import { PullRequest } from '../Models/index';
 import { Elements } from './Elements';
 
 export class PullRequestsElement {
     private readonly templateClass = 'pr-template';
 
-    private readonly modal: Element;
     private readonly template: Element;
 
     private onApproveHandler: (owner: string, name: string, number: number) => Promise<void>;
 
-    constructor(modal: Element) {
-        this.modal = modal;
+    constructor(private readonly modal: Element) {
         this.template = modal.querySelector('.pr-template');
         this.modal.addEventListener('hide.bs.modal', () => {
             this.unloadPullRequests();
